Initial refactor is better
