+++ src/slices/farming/Crop.ts
@@ -23,9 +23,9 @@ export class Crop {
   private getGrowthTime(cropType: string): number {
     // Define growth times for different crops
     const growthTimes: { [key: string]: number } = {
-      carrot: 5000,
-      raddish: 7000,
-      cauliflower: 6000,
+      carrot: 1200,
+      raddish: 1400,
+      cauliflower: 1600,
       // Add more crops as needed
     };
     return growthTimes[cropType] || 5000; // Default to 5000 ms
